Reconstruct TWRP only device tree for SM-G355H (kanas)
Reconstruction? I did have the tree but, as Life's christmas present to me was to
fail my hard disk, I lost all of it.

Trying to piece from my last working build, all I can do is something like this.

Hope this works. I used omni's source instead of the lineage one with the help of this guide
https://forum.xda-developers.com/showthread.php?p=32965365#post32965365

This is based on samsung jflte tree.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>